requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor
pymongo
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
passlib>=1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.1